set default model to schematics
