+++ basemodels/pydantic/manifest/data/groundtruth.py
@@ -1,20 +1,23 @@
-from typing import Union
-from schematics.models import Model
-from schematics.types import StringType, ListType, URLType, ModelType, IntType, FloatType, UnionType
+from typing_extensions import Literal
+from typing import Dict, Union
+from pydantic import BaseModel, HttpUrl, validate_model
+from typing import List, Optional
 
 
 def create_wrapper_model(type):
-    class WrapperModel(Model):
-        data = type
-
+    class WrapperModel(BaseModel):
+            data: Optional[type]
+            class Config:
+                    arbitrary_types_allowed = True
+    
     return WrapperModel
 
-
 def validate_wrapper_model(Model, data):
-    Model({"data": data}).validate()
+    *_, validation_error = validate_model(Model, {"data": data})
+    if validation_error:
+          raise validation_error
 
-
-groundtruth_entry_key_type = URLType()
+groundtruth_entry_key_type = HttpUrl
 GroundtruthEntryKeyModel = create_wrapper_model(groundtruth_entry_key_type)
 """
 Groundtruth file format for `image_label_binary` job type:
@@ -24,7 +27,7 @@ Groundtruth file format for `image_label_binary` job type:
   "https://domain.com/file2.jpeg": ["true", "true", "true"]
 }
 """
-ilb_groundtruth_entry_type = ListType(StringType(choices=["true", "false"]))
+ilb_groundtruth_entry_type = List[Literal["true", "false"]]
 ILBGroundtruthEntryModel = create_wrapper_model(ilb_groundtruth_entry_type)
 """
 Groundtruth file format for `image_label_multiple_choice` job type:
@@ -42,14 +45,14 @@ Groundtruth file format for `image_label_multiple_choice` job type:
   ]
 }
 """
-ilmc_groundtruth_entry_type = ListType(ListType(StringType))
+ilmc_groundtruth_entry_type = List[List[str]]
 ILMCGroundtruthEntryModel = create_wrapper_model(ilmc_groundtruth_entry_type)
 
 
-class ILASGroundtruthEntry(Model):
-    entity_name = UnionType([IntType, FloatType])
-    entity_type = StringType()
-    entity_coords = ListType(UnionType([IntType, FloatType]))
+class ILASGroundtruthEntry(BaseModel):
+    entity_name: Union[int, float]
+    entity_type: str
+    entity_coords: List[Union[int, float]]
 
 
 """
@@ -67,7 +70,7 @@ Groundtruth file format for `image_label_area_select` job type
   ]
 }
 """
-ilas_groundtruth_entry_type = ListType(ListType(ModelType(ILASGroundtruthEntry)))
+ilas_groundtruth_entry_type = List[List[ILASGroundtruthEntry]]
 ILASGroundtruthEntryModel = create_wrapper_model(ilas_groundtruth_entry_type)
 
 groundtruth_entry_models_map = {
